refactor: `CardUiModel` 패키지 이동 및 적용
